blomotobackend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
pillow==11.1.0
